pykmip-restapi/routes/kmip_routes.py

from flask import Blueprint, request, jsonify

class KMIPRoutes:
    def __init__(self, kmip_service):
        self.kmip_service = kmip_service
        self.bp = Blueprint('kmip', __name__)
        self._register_routes()

    def _register_routes(self):
        self.bp.route('/get_barbican_id', methods=['GET'])(self.get_barbican_id)
        self.bp.route('/update_policy', methods=['POST'])(self.update_policy)
        self.bp.route('/kmip_register', methods=['POST'])(self.register_kmip)

    def get_barbican_id(self):
        kmip_id = request.args.get('kmip_id')
        if not kmip_id:
            return jsonify({"error": "Missing kmip_id"}), 400
        return jsonify(self.kmip_service.execute_mysql_queries(kmip_id))

    def update_policy(self):
        data = request.get_json()
        return jsonify({"message": "Update policy logic goes here"})  # Placeholder

    def register_kmip(self):
        data = request.get_json()
        url = data.get('url')
        owner = data.get('owner')
        policy = data.get('policy')
        if not url or not owner or not policy:
            return jsonify({"error": "Missing parameters"}), 400
        return jsonify(self.kmip_service.register_kmip_object(url, owner, policy))
